Moved expo_to_hdf5 in commonlib
