+++ openquake/commonlib/tests/data/grm_exposure.csv
