+++ openquake/commonlib/tests/data/grm_exposure.xml
