+++ utils/build_global_exposure
@@ -23,10 +23,10 @@ import pandas
 import numpy
 from openquake.baselib import sap, performance
 from openquake.hazardlib import nrml
+from openquake.risklib.riskmodels import CompositeRiskModel, RiskFuncList
 from openquake.commonlib.datastore import build_dstore_log
 from openquake.commonlib.oqvalidation import OqParam
-from openquake.risklib.riskmodels import CompositeRiskModel, RiskFuncList
-from openquake.risklib import expo_to_hdf5
+from openquake.commonlib import expo_to_hdf5
 from openquake.qa_tests_data import mosaic
 
 U16 = numpy.uint16
